medical-agent/requirements.txt
livekit>=0.11.0
livekit-agents>=0.8.0
livekit-plugins-silero>=0.5.0
livekit-plugins-cartesia>=0.5.0
livekit-plugins-openai>=0.5.0
livekit-plugins-turn-detector>=0.1.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
